requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
pymysql
